[add update]Favorite.jsのファイル位置変更 コメント数表示のため、FavoriteCard.jsコンポーネント作成 マイページのお気に入り画面にコメント数を表示 Pages.jsのファイル位置変更
